点击关闭按钮时调用kernel退出接口

+++ electron/src/main.js
@@ -101,10 +101,9 @@ app.whenReady().then(() => {
 
 app.on('window-all-closed', () => {
     if (process.platform !== 'darwin') {
-        if (kernelProcess.exitCode === null) {
+        if (kernelProcess && kernelProcess.exitCode === null) {
             kernelProcess.kill()
         }
         app.quit();
-
     }
 });
